aegis/requirements.txt
# Core — pin loosely for compatibility across platforms
torch>=2.0.0,<3.0.0
torch-geometric>=2.4.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
PyYAML>=6.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0
matplotlib>=3.7.0
pytest>=7.4.0
streamlit>=1.28.0
openai>=1.0.0
